RAG_Demo/Knowledge_rag/requirements.txt
sentence-transformers
faiss-cpu
neo4j
fastapi
uvicorn
numpy